Extract diversion operators unit tests to another file
